+++ src/kard/users/attributions/types/offer_medium.py
@@ -2,4 +2,4 @@
 
 import typing
 
-OfferMedium = typing.Union[typing.Literal["BROWSE", "EMAIL", "MAP", "SEARCH"], typing.Any]
+OfferMedium = typing.Union[typing.Literal["BROWSE", "EMAIL", "MAP", "SEARCH", "CTA"], typing.Any]